requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
autopep8==2.3.1
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.110.3
fastavro==1.9.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.7
langchain-cli==0.0.25
langchain-cohere==0.1.9
langchain-community==0.2.7
langchain-core==0.2.18
langchain-experimental==0.0.62
langchain-openai==0.1.16
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.2
langserve==0.2.2
langsmith==0.1.85
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.35.14
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
pinecone==4.0.0
pinecone-client==3.2.2
protobuf==5.27.2
pyarrow==16.1.0
pycodestyle==2.12.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.3.0
pyproject-toml==0.0.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.22.0
sentence-transformers==3.0.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
sse-starlette==1.8.2
starlette==0.37.2
streamlit==1.37.1
sympy==1.13.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.9.4
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.23.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4